armeabi-v7a and x86 libraries now build on Mac
I will have to try on Linux when I get back to my Linux machine.
The GLIB and ICU libraries build fine on armeabi, armeabi-v7a, x86, mips, x86_64, and arm64-v8a
mips64 fails to even configure libiconv; have to dive in
JavaScriptCore fails on mips due to some missing assembly code (may be an out-of-sync problem with webkit)
JavaScriptCore fails on armeabi at the very, very end when it tries to assemble incompatible ARM code
JavaScriptCore fails on x86_64 and arm64-v8a due to what appears to be a GCC 4.9 compiler bug

Still working on the above.