Fixed issues with API < 19.  Now runs on Jellybean (API 16) and newer.
Due to a small stack on Android's main thread, some large operations may cause
a stack overflow.  To address this, all heavy or VM-modifying operations now
run in a separate worker thread with a large stack.  This issue always existed,
but was exarcerbated by an even smaller stack on older APIs.

Now all operations are put into a serial dispatch queue, that (1) guarantees
a large stack (default is 1MB), and (2) ensures thread safety as an added
benefit.